Adding Sprites & code
Adding the images of the microbes to be used in the body tracking script for OMSI.

Added script to upload the images and have them moves around the screen in a sine wave form until the camera detects a human, then the microbes start to flock to the skeleton that the camera is tracking. When the skeleton leaves the frame, the microbes start to swing around the screen again freely and independently from one another.